+++ app/auth/views/change_email.py
@@ -3,6 +3,7 @@ from flask_login import login_user
 
 from app.auth.base import auth_bp
 from app.db import Session
+from app.log import LOG
 from app.models import EmailChange, ResetPasswordCode
 
 
@@ -22,12 +23,14 @@ def change_email():
         return render_template("auth/change_email.html")
 
     user = email_change.user
+    old_email = user.email
     user.email = email_change.new_email
 
     EmailChange.delete(email_change.id)
     ResetPasswordCode.filter_by(user_id=user.id).delete()
     Session.commit()
 
+    LOG.i(f"User {user} has changed their email from {old_email} to {user.email}")
     flash("Your new email has been updated", "success")
 
     login_user(user)
