+++ templates/header.html
@@ -148,6 +148,10 @@
             <a class="dropdown-item mb-3" href="{{ url_for('dashboard.api_key') }}">
               <i class="dropdown-icon fa fa-key"></i> API Keys
             </a>
+            <a class="dropdown-item mb-3"
+               href="{{ url_for('dashboard.account_setting') }}">
+              <i class="dropdown-icon fa fa-user"></i> Account settings
+            </a>
             <a class="dropdown-item" href="{{ url_for('auth.logout') }}">
               <i class="dropdown-icon fe fe-log-out"></i> Sign out
             </a>
