+++ tests/dashboard/test_sudo_setting.py
